CPR-603 S3 code for large messages is in SNSCLientFactory
